Move detail page related components to "Detail" folder
