+++ src/components/Detail/Annotations/AnnotationItemContent.tsx
@@ -1,7 +1,10 @@
 import React from "react";
 import { Button } from "react-aria-components";
-import { AnnoRepoAnnotation } from "../../model/AnnoRepoAnnotation";
-import { projectConfigSelector, useProjectStore } from "../../stores/project";
+import { AnnoRepoAnnotation } from "../../../model/AnnoRepoAnnotation";
+import {
+  projectConfigSelector,
+  useProjectStore,
+} from "../../../stores/project";
 
 type AnnotationContentProps = {
   annotation: AnnoRepoAnnotation;
