+++ src/components/Detail/Footer/BrowseScanButtons.tsx
