+++ src/components/Detail/Footer/DetailSearchResultsNavigation.tsx
