+++ src/components/Detail/Footer/Footer.tsx
