+++ src/components/Detail/Footer/FooterLink.tsx
