+++ src/components/Detail/Mirador/Mirador.tsx
