+++ src/components/Detail/Mirador/defaultMiradorConfig.ts
