+++ src/components/Detail/Mirador/utils/observeMiradorStore.ts
