+++ src/components/Detail/Mirador/zoomAnnoMirador.ts
