+++ src/components/Detail/Text/Annotated/AnnotatedText.tsx
