+++ src/components/Detail/Text/Annotated/AnnotationModel.ts
