+++ src/components/Detail/Text/Annotated/LineSegment.tsx
