+++ src/components/Detail/Text/Annotated/LineSegmentWithAnnotations.tsx
