+++ src/components/Detail/Text/Annotated/LineSegmentsViewer.tsx
