+++ src/components/Detail/Text/Annotated/MetadataDetailLabelValues.tsx
