+++ src/components/Detail/Text/Annotated/SegmentGroup.tsx
