+++ src/components/Detail/Text/Annotated/SegmentedLine.tsx
