+++ src/components/Detail/Text/Annotated/SpanButton.tsx
