+++ src/components/Detail/Text/Annotated/annotated.css
