+++ src/components/Detail/Text/Annotated/details/EntitySummaryProps.ts
