+++ src/components/Detail/Text/Annotated/test/resources/dummyLogicalTextAnnotations.ts
