+++ src/components/Detail/Text/Annotated/utils/AnnotationSegmenter.spec.ts
