+++ src/components/Detail/Text/Annotated/utils/AnnotationSegmenter.ts
