+++ src/components/Detail/Text/Annotated/utils/createLineOffsets.spec.ts
