+++ src/components/Detail/Text/Annotated/utils/createLineOffsets.ts
