+++ src/components/Detail/Text/Annotated/utils/createLineSearchOffsets.spec.ts
