+++ src/components/Detail/Text/Annotated/utils/createSearchHighlightOffsets.ts
