+++ src/components/Detail/Text/Annotated/utils/detailPath.ts
