+++ src/components/Detail/Text/Annotated/utils/getAnnotationsByTypes.ts
