+++ src/components/Detail/Text/Annotated/utils/groupSegmentsByGroupId.ts
