+++ src/components/Detail/Text/Annotated/utils/listAnnotationOffsets.spec.ts
