+++ src/components/Detail/Text/Annotated/utils/listOffsetsByChar.ts
