+++ src/components/Detail/Text/Annotated/utils/trimMiddle.ts
