+++ src/components/Detail/Text/BroccoliViewPosition.ts
