+++ src/components/Detail/Text/StyledText.tsx
