+++ src/components/Detail/Text/TextComponent.tsx
