+++ src/components/Detail/Text/TextPanel.tsx
