+++ src/components/Detail/Text/TextPanels.tsx
