+++ src/components/Detail/Text/ToggleTextPanels.tsx
