+++ src/components/Detail/Text/createSearchRegex.tsx
